python/random_cluster/requirements.txt
biopython==1.78
numpy==1.19
